src/config.py
import os
from dotenv import load_dotenv

load_dotenv()

DATABASE_URL = os.getenv("DATABASE_URL")
SECRET_KEY = os.getenv("SECRET_KEY")
MONITORING_API_KEY = os.getenv("MONITORING_API_KEY")
ALGORITHM = "HS256"